refactor: rename analysis workflow builder module
