+++ analysis/application/analysis_policy_facade.py
@@ -1,5 +1,5 @@
 from ...activities.application.activity_selection_state import ActivitySelectionState
-from .analysis_request_building import build_analysis_workflow
+from .analysis_workflow_building import build_analysis_workflow
 from .analysis_workflow_execution import run_analysis_workflow
